fix: auto-deploy node scripts to Windmill on startup
The http-call script in Windmill was stale (pre-March 3) and missing
identity header injection (x-org-id, x-user-id, x-run-id). This caused
all campaign workflow runs to fail with 400 "expected string, received
undefined" because downstream services require these headers.

- Move deployNodes logic to src/lib/deploy-nodes.ts (inside rootDir)
- Call deployNodes() on startup before accepting traffic
- Crash on failure to prevent running with stale scripts
- scripts/deploy-nodes.ts becomes a thin CLI wrapper
- Add regression test for gate-check identity header injection
